Vehicle filters

+++ Rentacar.API/Rentacar.Dto/Enums/TransmissionTypeEnum.cs
@@ -1,6 +1,6 @@
 ﻿namespace Rentacar.Dto.Enums
 {
-    public enum TransmissionTypeEnum : int
+    public enum TransmissionTypeEnum : short
     {
         Manual = 1,
         Automatic = 2
